--- src/test/java/io/swagger/codegen/languages/DefaultCodegenConfigTest.java
@@ -1,110 +1,0 @@
-package io.swagger.codegen.languages;
-
-import io.swagger.codegen.CodegenArgument;
-import io.swagger.codegen.CodegenConstants;
-import io.swagger.codegen.CodegenProperty;
-import io.swagger.codegen.CodegenType;
-import io.swagger.codegen.v3.generators.DefaultCodegenConfig;
-import io.swagger.v3.oas.models.media.NumberSchema;
-import io.swagger.v3.oas.models.media.Schema;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import java.math.BigDecimal;
-import java.util.List;
-
-public class DefaultCodegenConfigTest {
-
-    @Test
-    public void testInitialValues() throws Exception {
-        final DefaultCodegenConfig codegen = new P_DefaultCodegenConfig();
-        codegen.processOpts();
-
-        Assert.assertEquals(codegen.modelPackage, "");
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.MODEL_PACKAGE), null);
-        Assert.assertEquals(codegen.apiPackage, "");
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.API_PACKAGE), null);
-        Assert.assertEquals(codegen.sortParamsByRequiredFlag, Boolean.TRUE);
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.SORT_PARAMS_BY_REQUIRED_FLAG), Boolean.TRUE);
-        Assert.assertEquals(codegen.hideGenerationTimestamp, Boolean.TRUE);
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.TRUE);
-    }
-
-    @Test
-    public void testSetters() throws Exception {
-        final DefaultCodegenConfig codegen = new P_DefaultCodegenConfig();
-        codegen.setModelPackage("xxx.yyyyy.zzzzzzz.model");
-        codegen.setApiPackage("xxx.yyyyy.zzzzzzz.api");
-        codegen.setSortParamsByRequiredFlag(false);
-        codegen.setHideGenerationTimestamp(false);
-        codegen.processOpts();
-
-        Assert.assertEquals(codegen.modelPackage, "xxx.yyyyy.zzzzzzz.model");
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.MODEL_PACKAGE), "xxx.yyyyy.zzzzzzz.model");
-        Assert.assertEquals(codegen.apiPackage, "xxx.yyyyy.zzzzzzz.api");
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.API_PACKAGE), "xxx.yyyyy.zzzzzzz.api");
-        Assert.assertEquals(codegen.sortParamsByRequiredFlag, Boolean.FALSE);
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.SORT_PARAMS_BY_REQUIRED_FLAG), Boolean.FALSE);
-        Assert.assertEquals(codegen.hideGenerationTimestamp, Boolean.FALSE);
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.FALSE);
-    }
-
-    @Test
-    public void testPutAdditionalProperties() throws Exception {
-        final DefaultCodegenConfig codegen = new P_DefaultCodegenConfig();
-        codegen.additionalProperties().put(CodegenConstants.MODEL_PACKAGE, "xx.yyyyy.model");
-        codegen.additionalProperties().put(CodegenConstants.API_PACKAGE, "xx.yyyyy.api");
-        codegen.additionalProperties().put(CodegenConstants.SORT_PARAMS_BY_REQUIRED_FLAG, false);
-        codegen.additionalProperties().put(CodegenConstants.HIDE_GENERATION_TIMESTAMP, false);
-        codegen.processOpts();
-
-        Assert.assertEquals(codegen.modelPackage, "xx.yyyyy.model");
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.MODEL_PACKAGE), "xx.yyyyy.model");
-        Assert.assertEquals(codegen.apiPackage, "xx.yyyyy.api");
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.API_PACKAGE), "xx.yyyyy.api");
-        Assert.assertEquals(codegen.sortParamsByRequiredFlag, Boolean.FALSE);
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.SORT_PARAMS_BY_REQUIRED_FLAG), Boolean.FALSE);
-        Assert.assertEquals(codegen.hideGenerationTimestamp, Boolean.FALSE);
-        Assert.assertEquals(codegen.additionalProperties().get(CodegenConstants.HIDE_GENERATION_TIMESTAMP), Boolean.FALSE);
-    }
-
-    @Test
-    public void testNumberSchemaMinMax() {
-        Schema schema = new NumberSchema()
-                .minimum(BigDecimal.valueOf(50))
-                .maximum(BigDecimal.valueOf(1000));
-
-        final DefaultCodegenConfig codegen = new P_DefaultCodegenConfig();
-        CodegenProperty codegenProperty = codegen.fromProperty("test", schema);
-
-        Assert.assertEquals(codegenProperty.minimum, "50");
-        Assert.assertEquals(codegenProperty.maximum, "1000");
-    }
-
-    private static class P_DefaultCodegenConfig extends DefaultCodegenConfig{
-        @Override
-        public String getArgumentsLocation() {
-            return null;
-        }
-
-        @Override
-        public CodegenType getTag() {
-            return null;
-        }
-
-        @Override
-        public String getName() {
-            return null;
-        }
-
-        @Override
-        public String getHelp() {
-            return null;
-        }
-
-        @Override
-        public List<CodegenArgument> readLanguageArguments() {
-            return null;
-        }
-    }
-}
