--- src/test/java/io/swagger/codegen/options/JavaInflectorServerOptionsProvider.java
@@ -1,13 +1,0 @@
-package io.swagger.codegen.options;
-
-public class JavaInflectorServerOptionsProvider extends JavaOptionsProvider {
-    @Override
-    public String getLanguage() {
-        return "inflector";
-    }
-
-    @Override
-    public boolean isServer() {
-        return true;
-    }
-}
