--- src/test/java/io/swagger/codegen/options/OptionsProvider.java
@@ -1,10 +1,0 @@
-package io.swagger.codegen.options;
-
-import java.util.Map;
-
-public interface OptionsProvider {
-
-    String getLanguage();
-    Map<String, String> createOptions();
-    boolean isServer();
-}
