+++ src/test/java/io/swagger/codegen/v3/generators/options/OptionsProvider.java
@@ -1,0 +1,10 @@
+package io.swagger.codegen.v3.generators.options;
+
+import java.util.Map;
+
+public interface OptionsProvider {
+
+    String getLanguage();
+    Map<String, String> createOptions();
+    boolean isServer();
+}
