+++ src/test/java/io/swagger/codegen/v3/generators/typescript/angular/TypeScriptAngularClientOptionsTest.java
@@ -1,0 +1,36 @@
+package io.swagger.codegen.v3.generators.typescript.angular;
+
+import io.swagger.codegen.languages.TypeScriptAngularClientCodegen;
+import io.swagger.codegen.v3.CodegenConfig;
+import io.swagger.codegen.v3.generators.AbstractOptionsTest;
+import io.swagger.codegen.v3.generators.options.TypeScriptAngularClientOptionsProvider;
+import mockit.Expectations;
+import mockit.Tested;
+
+public class TypeScriptAngularClientOptionsTest extends AbstractOptionsTest {
+
+    @Tested
+    private TypeScriptAngularClientCodegen clientCodegen;
+
+    public TypeScriptAngularClientOptionsTest() {
+        super(new TypeScriptAngularClientOptionsProvider());
+    }
+
+    @Override
+    protected CodegenConfig getCodegenConfig() {
+        return clientCodegen;
+    }
+
+    @SuppressWarnings("unused")
+    @Override
+    protected void setExpectations() {
+        new Expectations(clientCodegen) {{
+            clientCodegen.setSortParamsByRequiredFlag(Boolean.valueOf(TypeScriptAngularClientOptionsProvider.SORT_PARAMS_VALUE));
+            times = 1;
+            clientCodegen.setModelPropertyNaming(TypeScriptAngularClientOptionsProvider.MODEL_PROPERTY_NAMING_VALUE);
+            times = 1;
+            clientCodegen.setSupportsES6(Boolean.valueOf(TypeScriptAngularClientOptionsProvider.SUPPORTS_ES6_VALUE));
+            times = 1;
+        }};
+    }
+}
